@@ -1,14 +1,13 @@
 package com.github.nukc.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 
 import com.github.nukc.stateview.StateView;
 import com.github.nukc.stateview.animations.FadeScaleAnimatorProvider;
-import com.github.nukc.stateview.animations.SlideAnimatorProvider;
 
 public class WrapActivity extends AppCompatActivity {
 
