@@ -1,10 +1,11 @@
 package com.github.nukc.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.github.nukc.stateview.StateView;
 import com.github.nukc.stateview.animations.SlideAnimatorProvider;
@@ -18,7 +19,7 @@ public class ConstraintLayoutActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_constraint_layout);
 
-        mStateView = findViewById(R.id.stateView);
+        mStateView = StateView.inject(findViewById(R.id.text));
         mStateView.setAnimatorProvider(new SlideAnimatorProvider());
         mStateView.setEmptyResource(R.layout.view_empty);
         mStateView.setRetryResource(R.layout.view_retry);
