@@ -1,9 +1,10 @@
 package com.github.nukc.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.github.nukc.stateview.StateView;
 
