@@ -1,9 +1,10 @@
 package com.github.nukc.stateview;
 
-import android.support.constraint.ConstraintLayout;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
+
+import androidx.constraintlayout.widget.ConstraintLayout;
 
 class Injector {
 
